+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## [3.2.6]
+## [patch]
 
 - Added `extract` function to the `Schnorr256` library to extract a secret from a standard/adaptor Schnorr signature pair.