+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [patch]
 
 - Added `ED256` library for 256-bit Twisted Edwards curve arithmetic.
+- Optimized `U512` library `eq(mload(a_), 0)` -> `iszero(mload(a_))`.
 
 ## [3.2.7]